Module asfa::cli

Modules

color

text

Structs

Opts

WaitingSpinner

Spinner that spins until finish() is called.

Enums

SpinnerSetting

UserCommand

Functions

draw_boxed

join_frames

Replace portions in raw that are a horizontal line (‘─’) with joiner where content contains a vertical line (‘│’). This makes it possible to join frames.

spinner

style_progress_bar_count

Progress bar style for counts

style_progress_bar_transfer

Progress bar style for file transfers

style_spinner