Struct asfa::cmd::check::Check

pub struct Check {
    files: Vec<PathBuf>,
    details: bool,
    filenames: bool,
    no_details: bool,
    url_only: bool,
    with_time: bool,
    with_size: bool,
}

Check if a given local file is already present on the remote site.

Fields
files: Vec<PathBuf>

File(s) to check for.

details: bool

Show all details, can be set globally in config file.

filenames: bool

Show no full urls but rather filenames only. Makes for more concise output.

no_details: bool

If details is set to true in config, –no-details can be specified to suppress output.

url_only: bool

Only list the remote URLs (useful for copying and scripting).

with_time: bool

Print remote modification time

with_size: bool

Print file sizes