Struct asfa::cmd::clean::Clean

pub struct Clean {

    all: bool,
    details: bool,
    files: Vec<String>,
    filter: Option<String>,
    last: Option<usize>,
    indices: Vec<i64>,
    no_confirm: bool,
    no_details: bool,
    select_newer: Option<String>,
    select_older: Option<String>,
    sort_size: bool,
    sort_time: bool,
    reverse: bool,

}

Clear already uploaded files.

Fields
all: bool

Clean all remote files (dangerous!)

details: bool

Show all details in confirmation, can be set globally in config file.

files: Vec<String>

Explicit file to delete

filter: Option<String>

Filter filenames by regex. See https://docs.rs/regex/latest/regex/#syntax

last: Option<usize>

Delete last

indices: Vec<i64>

Indices of files to delete as returned by list command.

no_confirm: bool

Disable confirming deletions

no_details: bool

If details is set to true in config, –no-details can be specified to suppress output.

select_newer: Option<String>

Select files newer than the given duration. Durations can be:seconds (sec, s), minutes (min, m), days (d), weeks (w), months (M) or years (y).

select_older: Option<String>

Select files older than the given duration. Durations can be:seconds (sec, s), minutes (min, m), days (d), weeks (w), months (M) or years (y).

sort_size: bool

Sort by size (useful when specifying --filter/--last)

sort_time: bool

Sort by modification time (useful when using --filter and --last).

reverse: bool

Reverse ordering (useful when specifying --last and --sort-{size,time})

Implementations

impl Clean

fn user_confirm_deletion(&self, files: &FileListing<'_>) -> Result<bool>

Have the user confirm deletions