Struct asfa::cmd::push::Push

pub struct Push {
    alias: Vec<String>,
    expire: Option<String>,
    files: Vec<PathBuf>,
    limit_mbits: Option<f64>,
    limit_kbytes: Option<f64>,
    prefix: Option<String>,
    suffix: Option<String>,
}

Upload new files.

Fields
alias: Vec<String>

Alias/file name on the remote site.

If you specify multiple files to upload you can either specify no aliases or as many aliases as there are files to upload.

expire: Option<String>

Expire the uploaded file after the given amount of time via at-scheduled remote job.

Select files newer than the given duration. Durations can be: seconds (sec, s), minutes (min, m), days (d), weeks (w), months (M) or years (y).

Mininum time till expiration is a minute.

Any setting specified via command line overwrites settings from config files.

A globally set expiration setting can overwritten by specifying “none”.

files: Vec<PathBuf>

File(s) to upload.

limit_mbits: Option<f64>

Limit upload speed (in Mbit/s). Please note that the upload speed will be shown in {M,K}Bytes/s, but most internet providers specify upload speeds in Mbits/s. This option makes it easier to specify what portion of your available upload speed to use. See also: –limit-kbytes

limit_kbytes: Option<f64>

Limit upload speed (in kByte/s).

prefix: Option<String>

Upload all files with the given prefix prepended. This is especially useful to give a bunch of files with generic names (e.g., plots) more context.

Example: --prefix foo_ causes bar.png to be uploaded as foo_bar.png.

suffix: Option<String>

Upload all files with the given suffix appended while not altering the file extension. This is especially useful to give a bunch of files with generic names (e.g., plots) more context.

NOTE: Only the last extension is honored.

Example: --suffix _bar causes foo.png to be uploaded as foo_bar.png.

Implementations

impl Push

fn upload(
&self,
session: &SshSession<'_>,
config: &Config,
to_upload: &Path,
target_name: &str
) -> Result<()>

fn transform_filename(&self, file: &Path) -> Result<String>