Struct asfa::cfg::Host

pub struct Host {
    pub alias: String,
    pub auth: Auth,
    pub expire: Option<String>,
    pub folder: PathBuf,
    pub group: Option<String>,
    pub hostname: Option<String>,
    pub password: Option<String>,
    pub prefix_length: u8,
    pub url: String,
    pub user: Option<String>,
}

A host entry

Fields
alias: String

Alias under which the host is known

auth: Auth

Overwrite global authentication settings for this host.

expire: Option<String>

Expire the uploaded file after the given amount of time via at-scheduled remote job.

Select files newer than the given duration. Durations can be: seconds (sec, s), minutes (min, m), days (d), weeks (w), months (M) or years (y).

Mininum time till expiration is a minute.

Overrides the global setting.

folder: PathBuf

In which folder do we store files on the host.

group: Option<String>

In case files on the remote site need to have a special group setting in order to be readable by the webserver.

hostname: Option<String>

Self-explanatory (if not set alias will be used)

password: Option<String>

If the user REALLY REALLY wants to, a plaintext password can be provided (but it is not recommended!).

prefix_length: u8

Length of prefix to use

url: String

url-prefix to apply to file link

user: Option<String>

The user to sign in, otherwise ssh config will be used.

Implementations

impl Host

fn from_yaml(alias: String, input: &Yaml) -> Result<Host>

fn from_yaml_with_config(
alias: String,
input: &Yaml,
config: &Config
) -> Result<Host>

pub fn get_url(&self, file: &str) -> Result<String>

Get URL to given destination for this host.

Prepends url and performs character escapes.