Struct asfa::file_listing::FileListing

pub struct FileListing<'a> {
    pub num_files: usize,
    all_files: HashMap<usize, PathBuf>,
    pub indices: Vec<usize>,
    pub stats: Option<HashMap<usize, FileStat>>,
    ssh: &'a SshSession<'a>,
}

Helper structure to avoid re-implementing file listing capabilities for all commands.

Fields
num_files: usizeall_files: HashMap<usize, PathBuf>indices: Vec<usize>stats: Option<HashMap<usize, FileStat>>ssh: &'a SshSession<'a>
Implementations

impl<'a> FileListing<'a>

pub fn new(ssh: &'a SshSession<'_>) -> Result<FileListing<'a>>

pub fn by_filter(self, filter: Option<&str>) -> Result<Self>

Select all files which name matches regex

pub fn by_indices(self, indices: &[i64]) -> Result<Self>

Select all files with corresponding indices

pub fn by_hash<T: AsRef<str>>( self, names: impl IntoIterator<Item = T>, prefix_length: u8, bail_when_missing: bool ) -> Result<Self>

Select all files that have the same hash as the names given

pub fn count(&self) -> usize

Return count of currently selected files

pub fn has_stats(&self) -> bool

Check if file listing has stats

pub fn first(self, n: Option<usize>) -> Self

Only use first n files

pub fn iter(&'a self) -> FileListingIter<'a>

pub fn last(self, n: Option<usize>) -> Self

Only use last n files

pub fn format_files( &self, host: Option<&Host>, filename_only: bool, with_size: bool, with_time: bool ) -> Result<Vec<String>>

Get formatted lines to be printed with draw_boxed() if stdout is a tty.

If stdout is no tty, simply format contents separated by tabs for easy parsing.

If filename_only is specified, prints url if host is specified, otherwise the relative path.

pub fn revert(self, do_revert: bool) -> Self

pub fn select_newer(self, user_duration: Option<&str>) -> Result<Self>

pub fn select_older(self, user_duration: Option<&str>) -> Result<Self>

pub fn sort_by_size(self, sort_by_size: bool) -> Result<Self>

pub fn sort_by_time(self, sort_by_time: bool) -> Result<Self>

pub fn with_all(self, select_all: bool) -> Self

Simply select all files if argument is true

pub fn with_all_if_none(self, doit: bool) -> Self

Add all if, so far, no files have been selected and the boolean switch is set.

pub fn with_stats(self, with_stats: bool) -> Result<Self>

fn column_size(&self, stat: &FileStat) -> Result<String>

fn column_time(&self, stat: &FileStat) -> Result<String>

fn ensure_stats(&mut self) -> Result<()>

fn make_unique<I: IntoIterator<Item = usize>>(indices: I) -> Vec<usize>

fn get_num_digits(&self) -> usize

Get number of digits for index

fn get_num_digits_rev(&self) -> Result<usize>

Get number of digits for reverse index

fn filter_by_time( self, user_duration: Option<&str>, select_older: bool ) -> Result<Self>

Helper function that filters selected files by date.

If select_older == true, then only files older than user_duration will be kept; if false, only files newer than user_duration will be kept.