Struct asfa::ssh::SshSession

pub struct SshSession<'a> {
    raw: Session,
    pub host: &'a Host,
    cfg_openssh: Option<OpenSshConfig>,
}

Fields
raw: Sessionhost: &'a Hostcfg_openssh: Option<OpenSshConfig>
Implementations

impl<'a> SshSession<'a>

pub fn adjust_group(&self, file: &Path, group: &str) -> Result<()>

Adjust the group of the remote target recursively

pub fn all_files(&self) -> Result<Vec<PathBuf>>

List all files present (relative to the current host’s base-folder).

fn auth(&self, auth: &Auth) -> Result<()>

Try all defined authentication methods in order

fn auth_agent(&self) -> Result<()>

fn auth_interactive(&self) -> Result<()>

fn auth_keyboard_interactive(&self) -> Result<()>

fn auth_private_key( &self, private_key_file: &str, private_key_file_password: Option<&str>, username: &str, interactive: bool ) -> Result<()>

fn auth_private_keys_openssh(&self)

Try to authenticate with private keys defined in openssh

Assumes that auth() already checked for supported pubkey authentication!

pub fn connect(host: &'a Host) -> Result<Self>

Create a new SSH session by connecting to the given host configuration.

First try authenticating with all agent identities then use an interactive password, if enabled.

pub fn exec_remote(&self, cmd: &str) -> Result<ExecutedRemoteCommand>

pub fn list_files(&self) -> Result<FileListing<'_>>

Get listing of files

pub fn make_folder(&self, path: &Path) -> Result<()>

Make folder on the remote site if it does not exist (relative to the current host’s base-folder).

pub fn mktemp(&self) -> Result<Tempfile<'_>>

Make remote file on remote side and return path to it.

pub fn get_auth_methods(&self) -> Result<HashSet<String>>

Get all available authentication methods

pub fn get_username(&self) -> String

Get username either from host, openSSH config or host system.

pub fn get_remote_hash(&self, path: &Path, length: u8) -> Result<String>

Get hash of the remote file (relative to the current host’s base-folder).

pub fn get_remote_hashes( &self, paths: &[&Path], length: u8 ) -> Result<Vec<String>>

Get hash of the remote file (relative to the current host’s base-folder).

pub fn prepend_base_folder(&self, path: &Path) -> PathBuf

pub fn remove_folder(&self, path: &Path) -> Result<()>

Remove the given folder and its contents (relative to the current host’s base-folder)

pub fn stat<'b, I: IntoIterator<Item = &'b Path> + Clone>( &self, paths: I ) -> Result<Vec<FileStat>>

Get stats about a remote files (relative to the current host’s base-folder)

pub fn stat_bulk<'b, I: IntoIterator<Item = &'b Path> + Clone>( &self, paths: I ) -> Result<Vec<FileStat>>

Get stats about a remote files (relative to the current host’s base-folder)

Faster version getting relevant information en bulk via find and xargs.

pub fn stat_fallback<'b, I: IntoIterator<Item = &'b Path>>( &self, paths: I ) -> Result<Vec<FileStat>>

Get stats about a remote files (relative to the current host’s base-folder)

Slower fallback that only relies on sftp functionality.

pub fn stat_single(&self, path: &Path) -> Result<FileStat>

Get stat for a single remote file (relative to base folder).

pub fn upload_file( &self, path_local: &Path, path_remote: &Path, limit_speed_bytes_per_second: Option<usize> ) -> Result<()>

Upload the given local path to the given remote path (relative to the current host’s base-folder)

fn stat_bulk_available(&self) -> Result<bool>

Check if necessary utilities for fast stat generation are available.