Module asfa::util

Macros

make_yaml_getter 🔒

Structs

InvalidYamlTypeError

Functions

get_array_from

get_bool_from

get_dict_from

get_explicit_hash 🔒

get_hash

Get hash digest of given file with chosen length

get_int_from

get_real_from

get_required

Raise an error if the required key does not exist.

get_string_from

yaml_string

Helper function static yaml strings.